% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\alias{store.collection}
\title{Store objects in a collection and read them back.}
\usage{
collection(name, create = FALSE, storage = filesystem(getwd()))

\method{store}{collection}(col, object, group = NULL, ...)
}
\arguments{
\item{col}{A \code{\link{collection}}.}

\item{group}{Store under this group name.}

\item{...}{Named tags (\code{store}) or additional parameters (\code{restore}).}

\item{obj}{Any object.}
}
\value{
Object identifier.
}
\description{
Store objects in a collection and read them back.
}

