% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sample_time_series}
\alias{sample_time_series}
\title{Produce a sample collection.}
\usage{
sample_time_series(n = 10, len = 84, seed = 1)
}
\arguments{
\item{n}{Number of objects to generate.}

\item{len}{Number of samples in each series.}

\item{seed}{Seed for the random number generator.}
}
\value{
A list of time series data frames.
}
\description{
Produce a sample collection.
}

