% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{filter}
\alias{filter}
\alias{filter.clist}
\alias{filter.collection}
\title{Filter objects in a collection or clist.}
\usage{
filter.collection(col, ...)

filter.clist(col, ...)
}
\arguments{
\item{col}{Collection or a \code{clist}.}

\item{col}{Collection or a \code{clist}.}

\item{fun}{A \code{function} or a \code{deferred} fuction object.}
}
\value{
A \code{clist}.

A \code{list} of values returned from calls.
}
\description{
Filter objects in a collection or clist.

Apply a function to objects in a collection or clist.
}

