% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clist.R
\name{filter_.clist}
\alias{filter_.clist}
\title{Taps into \code{dplyr}'s filtering method.}
\usage{
\method{filter_}{clist}(.data, ..., .dots)
}
\arguments{
\item{.data}{A \code{clist} object.}

\item{...}{Filtering expression.}

\item{.dots}{A list of \code{lazyeval} expressions.}
}
\value{
A \code{clist} whose elements match the specified filter.
}
\description{
Taps into \code{dplyr}'s filtering method.
}

