% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{filter_.collection}
\alias{filter_.collection}
\title{Taps into \code{dplyr}'s filtering method.}
\usage{
filter_.collection(.data, ..., .dots)
}
\arguments{
\item{.data}{A \code{collection} object.}

\item{...}{Filtering expression.}

\item{.dots}{A list of \code{lazyeval} expressions.}
}
\value{
A \code{clist} whose elements match the specified filter.
}
\description{
Taps into \code{dplyr}'s filtering method.
}

