% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_code.R
\name{restore}
\alias{restore}
\title{Restore an object.}
\usage{
restore(from, id, ...)
}
\arguments{
\item{from}{Restore from collection, clist...}

\item{id}{Restore object by its identifier.}

\item{...}{Restore object(s) by their tags.}
}
\description{
Restore an object.
}

