% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sample_time_series}
\alias{sample_time_series}
\title{Produce a sample collection.}
\usage{
sample_time_series(n = 10, len = 84, dir = tempdir())
}
\arguments{
\item{n}{Number of objects to generate.}

\item{len}{Number of samples in each series.}

\item{dir}{Parent directory for the new collection.}
}
\value{
A \code{\link{collection}} object.
}
\description{
Produce a sample collection.
}

