% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{search_ids}
\alias{search_ids}
\title{Search for objects whose tags match filter(s).}
\usage{
search_ids(col, dots)
}
\arguments{
\item{col}{A \code{collection} object.}

\item{dots}{List of lazy-eval expressions.}
}
\value{
A vector of object identifiers.
}
\description{
Search for objects whose tags match filter(s).
}

