% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clist.R
\name{as_clist}
\alias{as_clist}
\alias{is_clist}
\alias{restore.clist}
\alias{store.clist}
\title{Create a \emph{collection list}.}
\usage{
as_clist(col)

is_clist(x)

\method{store}{clist}(clist, col)

\method{restore}{clist}(clist, .simplify = TRUE)
}
\arguments{
\item{col}{A \code{collection} or another \code{clist} (?)}

\item{x}{Object to be tested.}
}
\value{
\code{as_clist()} return an object of class \emph{clist}.
}
\description{
A \code{clist} is a container of references to objects stored in a
\code{collection}. When used with the \emph{verb} functions
(\code{filter}, \code{arrange}, \code{apply})

\code{is_clist()} tests whether an object is a \emph{clist}.

\code{store.clist()} stores the whole \emph{clist} in
collection \code{col}.
}

