% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{clist}
\alias{clist}
\title{Create a \emph{collection list}.}
\usage{
clist(col)
}
\arguments{
\item{col}{A \code{collection} or another \code{clist} (?)}
}
\value{
A \code{clist}.
}
\description{
A \code{clist} is a container of references to objects stored in a
\code{collection}. When used with the \emph{verb} functions
(\code{filter}, \code{arrange}, \code{apply})
}

