% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R, R/old_code.R
\name{collection}
\alias{add}
\alias{add_}
\alias{as_collection}
\alias{as_collection.character}
\alias{as_collection.default}
\alias{as_collection.object_store}
\alias{collection}
\alias{is_collection}
\alias{store.collection}
\title{Create a collection.}
\usage{
collection(..., storage)

is_collection(x)

as_collection(x, ...)

\method{as_collection}{default}(x, storage)

\method{as_collection}{object_store}(x, storage)

\method{as_collection}{character}(x, storage)

add(col, object, ...)

add_(col, object, tags, auto_tags = FALSE)

\method{store}{collection}(col, object, group = NULL, ...)
}
\arguments{
\item{...}{All values are concatenated and \code{unlist}ed into
a \code{character} vector of identifiers that must be present
in the \code{storage}.}

\item{storage}{A \code{\link[storage]{storage}} object where actual
objects are stored.}

\item{col}{A \code{\link{collection}}.}

\item{group}{Store under this group name.}

\item{obj}{Any object.}

\item{...}{Named tags (\code{store}) or additional parameters (\code{restore}).}
}
\value{
Object identifier.
}
\description{
Create a collection.

In the \code{as_collection} for the \code{object_store}
             the \code{storage} parameter is ignored.

Store objects in a collection and read them back.
}

