% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sample_time_series}
\alias{sample_collection}
\alias{sample_time_series}
\title{Produce sample data or collection.}
\usage{
sample_time_series(n = 10, len = 84, seed = 1)

sample_collection(path = tempdir(), n = 10, len = 84, seed = 1)
}
\arguments{
\item{n}{Number of objects to generate.}

\item{len}{Number of samples in each series.}

\item{seed}{Seed for the random number generator.}

\item{path}{Where to create the collection.}
}
\value{
A list of time series data frames.

\code{sample_collection()} returns a \code{collection} object.
}
\description{
\code{sample_time_series()} generates a list of
\code{data.frame}s, each containing a sample time series. There are
four columns in each \code{data.frame}: columns \emph{a}, \emph{b}
and \emph{c} describe three independent variables and column \emph{x}
describes the dependent time series.

\code{sample_collection()} internally calls
\code{sample_time_series()} and stores its output in a file-based
collection.
}

