% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{store}
\alias{store}
\alias{store.default}
\alias{store.list}
\title{Store data in a collection.}
\usage{
store(x, y, group = NULL, ...)

\method{store}{default}(obj, collection, group = NULL, ...)

\method{store}{list}(obj, collection, group = NULL)
}
\arguments{
\item{x}{Object (contents) or collection (destination).}

\item{y}{Collection or object.}

\item{group}{Optional group name, especially useful when storing
multiple (related) objets.}

\item{...}{Optional tags.}

\item{obj}{Object to be store.}

\item{collection}{Collection to store \code{obj} in.}
}
\value{
Each \code{store()} method returns a \emph{clist}, which is
a handle to a set of objects in a \emph{collection}.
}
\description{
If the first argument is a collection, the \emph{store} operation is
performed in a single step - the whole object is stored under a single
identifier (key). If the first argument is the object, the \emph{store}
operation variant is chosen based on the object's class. E.g. a plain
\code{list} will be broken down into \code{n} \emph{store} operations,
one for each element of the list.

The default \code{store()} variant writes all data under
a single key (identifier).

The \code{list} variant of the \code{store()} method
stores each element of the list under a separate key (identifier).
This method does not take any additional, single-object-level tags
via \code{...}
}

